+++ react-app/src/components/music/index.js
@@ -4,3 +4,5 @@
 export { default as PlayerFooter } from './PlayerFooter';
 export { default as PlayerHeader } from './PlayerHeader';
 export { default as PlaylistSidebar } from './PlaylistSidebar';
+export { default as FullPlayerModal } from './FullPlayerModal';
+export { default as LyricsModal } from './LyricsModal';